+++ src/main/java/com/github/lipiridi/spotless/applier/action/ReformatFileAction.java
@@ -28,7 +28,7 @@ public class ReformatFileAction extends AnAction {
 
     @Override
     public @NotNull ActionUpdateThread getActionUpdateThread() {
-        return super.getActionUpdateThread();
+        return ActionUpdateThread.BGT;
     }
 
     @Override